+++ crates/o5/src/constants.rs
@@ -2,6 +2,7 @@
 
 use tor_llcrypto::pk::rsa::RSA_ID_LEN;
 
+pub use crate::common::ntor_arti::SESSION_ID_LEN;
 use crate::{
     common::{drbg, x25519_elligator2::REPRESENTATIVE_LENGTH},
     framing,
@@ -74,8 +75,6 @@ pub const MAX_CLOSE_DELAY_BYTES: usize = MAX_HANDSHAKE_LENGTH;
 
 pub const SEED_LENGTH: usize = drbg::SEED_LENGTH;
 pub const HEADER_LENGTH: usize = framing::FRAME_OVERHEAD + framing::MESSAGE_OVERHEAD;
-
-pub const SESSION_ID_LEN: usize = 8;
 
 pub const NODE_ID_LENGTH: usize = RSA_ID_LEN;
 pub const NODE_PUBKEY_LENGTH: usize = 32;
