+++ crates/o5/src/handshake/keys.rs
@@ -1,26 +1,28 @@
 use crate::{
     common::{
         mlkem1024_x25519::{PublicKey, StaticSecret},
-        ntor_arti::KeyGenerator,
+        ntor_arti::{KeyGenerator, SessionID, SessionIdentifier},
     },
+    constants::SESSION_ID_LEN,
+    framing::O5Codec,
     Result,
 };
 
 #[cfg(test)]
 use rand::{CryptoRng, RngCore};
-use subtle::Choice;
+use subtle::{Choice, ConstantTimeEq};
 use tor_bytes::SecretBuf;
 use tor_llcrypto::d::Shake256Reader;
-use tor_llcrypto::pk::ed25519::Ed25519Identity;
+use tor_llcrypto::pk::rsa::RsaIdentity;
 
 /// Key information about a relay used for the ntor v3 handshake.
 ///
 /// Contains a single curve25519 ntor onion key, and the relay's ed25519
 /// identity.
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct NtorV3PublicKey {
     /// The relay's identity.
-    pub(crate) id: Ed25519Identity,
+    pub(crate) id: RsaIdentity,
     /// The relay's onion key.
     pub(crate) pk: PublicKey,
 }
@@ -36,9 +38,12 @@ pub struct NtorV3SecretKey {
 impl NtorV3SecretKey {
     /// Construct a new NtorV3SecretKey from its components.
     #[allow(unused)]
-    pub(crate) fn new(sk: StaticSecret, pk: PublicKey, id: Ed25519Identity) -> Self {
+    pub(crate) fn new(sk: StaticSecret, id: RsaIdentity) -> Self {
         Self {
-            pk: NtorV3PublicKey { id, pk },
+            pk: NtorV3PublicKey {
+                id,
+                pk: PublicKey::from(&sk),
+            },
             sk,
         }
     }
@@ -46,7 +51,7 @@ impl NtorV3SecretKey {
     /// Generate a key using the given `rng`, suitable for testing.
     #[cfg(test)]
     pub(crate) fn generate_for_test<R: RngCore + CryptoRng>(rng: &mut R) -> Self {
-        let mut id = [0_u8; 32];
+        let mut id = [0_u8; 20];
         // Random bytes will work for testing, but aren't necessarily actually a valid id.
         rng.fill_bytes(&mut id);
 
@@ -62,16 +67,20 @@ impl NtorV3SecretKey {
     /// Checks whether `id` and `pk` match this secret key.
     ///
     /// Used to perform a constant-time secret key lookup.
-    pub(crate) fn matches(&self, id: Ed25519Identity, pk: PublicKey) -> Choice {
+    pub(crate) fn matches(&self, id: RsaIdentity, pk: PublicKey) -> Choice {
         id.as_bytes().ct_eq(self.pk.id.as_bytes()) & pk.as_bytes().ct_eq(self.pk.pk.as_bytes())
     }
 }
 
-/// A key generator returned from an ntor v3 handshake.
+pub trait NtorV3KeyGen: KeyGenerator + SessionIdentifier + Into<O5Codec> {}
+
+/// An instantiatable  key generator returned from an ntor v3 handshake.
 pub(crate) struct NtorV3KeyGenerator {
     /// The underlying `digest::XofReader`.
     pub(crate) reader: NtorV3XofReader,
 }
+
+impl NtorV3KeyGen for NtorV3KeyGenerator {}
 
 impl KeyGenerator for NtorV3KeyGenerator {
     fn expand(mut self, keylen: usize) -> Result<SecretBuf> {
@@ -88,5 +97,15 @@ pub(crate) struct NtorV3XofReader(pub(crate) Shake256Reader);
 impl digest::XofReader for NtorV3XofReader {
     fn read(&mut self, buffer: &mut [u8]) {
         self.0.read(buffer);
+    }
+}
+
+impl SessionIdentifier for NtorV3XofReader {
+    type ID = SessionID;
+
+    fn new_session_id(&mut self) -> Self::ID {
+        let mut s = [0u8; SESSION_ID_LEN];
+        <NtorV3XofReader as digest::XofReader>::read(self, &mut s);
+        SessionID::from(s)
     }
 }
